This update is to provide some clarification.
There has seen to be a lot of interest in this, and that is spectacular. It is also drawing a lot of requests of being incorporated to use as the main wallet to support more database support. I am only adding the line so that when someone comes across this project they understand what is going on prior to actually requesting help on other databases or insisting on immediate addition of a database.